requirements.txt
allpairspy==2.5.0
antlr4-python3-runtime==4.11.1
approvaltests==5.6.1
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
coverage==6.4.4
empty-files==0.0.3
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
Jinja2==3.1.2
lit==15.0.0.post0
MarkupSafe==2.1.1
mrjob==0.7.4
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pyperclip==1.8.2
pytest==7.1.3
PyYAML==6.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
zipp==3.8.1
